email-microservice/requirements.txt
Flask~=1.1.2
Jinja2~=2.11.3
MarkupSafe==1.1.1
itsdangerous==1.1.0
Werkzeug==1.0.1
Flask-Mail
python-dotenv